Progress 3
Added a game over method, that stops the main loop if the grid is full.
Changed dotsList into dotsDisplay, and instead added a playersDots matrix, which should keep track of the placed dots coordinates.
Added a Player class. which has some objects that could be needed later.
Added a setNames method in Player class, which checks checks for the game mode and sets the players' names accordingly.